backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
supabase==2.12.0
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.1.0
tensorflow==2.18.0
# torch==2.5.1  # Uncomment if using PyTorch instead of TensorFlow
scikit-learn==1.6.1
